Moves types in a separate file
