+++ src/hooks/network/index.ts
@@ -1,3 +1,4 @@
 import network from './network'
 
 export default network
+export * from './types'
